app/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
openai
httpx
python-multipart
pydantic-settings
google-generativeai
